refactor(Boot): US-803 create AppBootstrapper class"
Abstract system initialization logic into the App Bootstrapper class.
